+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Changed
 
+* `kwimage.Polygon.to_mask` will now pick dimensions to fit the polygon if they are unspecified.
 * `kwimage.Detections.to_coco` now handles the case where an entry in data or meta is None
 * Generic spatial list objects used by `kwimage.Detections` now implement the
   full MutableSequence API, and may explicitly inherit from it in the future.